worker/docking/requirements.txt
biopython
nglview
pandas
seaborn
vina